refactor: extract scratch buffers into dedicated microcrate
