+++ copybook-codec-memory/Cargo.toml
@@ -18,7 +18,7 @@ include = ["src/**", "Cargo.toml", "README.md", "LICENSE"]
 [dependencies]
 crossbeam-channel.workspace = true
 copybook-sequence-ring.workspace = true
-smallvec.workspace = true
+copybook-scratch.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
